boilerplate for dns query handle
